Split Assembly extensions
