+++ test/Tethos.Tests/Extensions/Assembly/AssemblyLoadingExtensionsTests.cs
@@ -1,0 +1,106 @@
+﻿namespace Tethos.Tests.Extensions.Assembly
+{
+    using System.Linq;
+    using System.Reflection;
+    using AutoFixture.Xunit2;
+    using FluentAssertions;
+    using Tethos.Extensions.Assembly;
+    using Xunit;
+
+    public class AssemblyLoadingExtensionsTests : BaseAutoMockingTest<AutoMockingContainer>
+    {
+        [Fact]
+        [Trait("Category", "Unit")]
+        public void TryToLoadAssembly_WithCorruptAssembly_ShouldReturnNull()
+        {
+            // Arrange
+            var corruptAssembly = "Fake.Core30.exe";
+
+            // Act
+            var actual = corruptAssembly.TryToLoadAssembly();
+
+            // Assert
+            actual.Should().BeNull();
+        }
+
+        [Theory]
+        [InlineData("Microsoft.Extensions.DependencyModel.dll")]
+        [InlineData("Fake.Framework461.exe")]
+        [InlineData("Fake.Core31.dll")]
+        [Trait("Category", "Unit")]
+        public void TryToLoadAssembly_ShouldLoadAssembly(string assemblyName)
+        {
+            // Arrange
+            var assembly = Assembly.LoadFrom(assemblyName);
+
+            // Act
+            var actual = assemblyName.TryToLoadAssembly();
+
+            // Assert
+            actual.Should().BeSameAs(assembly);
+        }
+
+        [Theory]
+        [AutoData]
+        [Trait("Category", "Unit")]
+        public void TryToLoadAssembly_UsingAssemblyName_ShouldReturnNull(string name)
+        {
+            // Arrange
+            var assemblyName = new AssemblyName(name);
+
+            // Act
+            var actual = assemblyName.TryToLoadAssembly();
+
+            // Assert
+            actual.Should().BeNull();
+        }
+
+        [Fact]
+        [Trait("Category", "Unit")]
+        public void TryToLoadAssembly_UsingAssemblyName_ShouldLoadAssembly()
+        {
+            // Arrange
+            var assembly = Assembly.GetExecutingAssembly();
+            var assemblyName = assembly.GetName();
+
+            // Act
+            var actual = assemblyName.TryToLoadAssembly();
+
+            // Assert
+            actual.Should().BeSameAs(assembly);
+        }
+
+        [Theory]
+        [InlineData("Tethos.dll", "Tethos.Tests.Common.dll")]
+        [InlineData("AutoFixture.dll", "FluentAssertions.dll", "Moq.dll", "xunit.core.dll")]
+        [InlineData("Castle.Core.dll", "Castle.Windsor.dll")]
+        [Trait("Category", "Unit")]
+        public void LoadAssemblies_ShouldLoad(params string[] assemblies)
+        {
+            // Arrange
+            var files = assemblies.Select(x => x.GetFile());
+            var expected = assemblies.Length;
+
+            // Act
+            var actual = files.LoadAssemblies();
+
+            // Assert
+            actual.Should().HaveCount(expected);
+        }
+
+        [Theory]
+        [InlineData("Fake.Core30.exe", "Fake.Core31.exe")]
+        [Trait("Category", "Unit")]
+        public void LoadAssemblies_ShouldSkip(params string[] assemblies)
+        {
+            // Arrange
+            var files = assemblies.Select(x => x.GetFile());
+
+            // Act
+            var actual = files.LoadAssemblies();
+
+            // Assert
+            actual.Should().BeEmpty();
+        }
+    }
+}
