+++ test/Tethos.Tests/Extensions/Assembly/AssemblyStub.cs
@@ -1,4 +1,4 @@
-﻿namespace Tethos.Tests
+﻿namespace Tethos.Tests.Extensions.Assembly
 {
     using System;
     using System.Reflection;
