+++ test/Tethos.Tests/Extensions/Assembly/FileExtensionsTests.cs
@@ -1,15 +1,15 @@
-﻿namespace Tethos.Tests.Extensions
+﻿namespace Tethos.Tests.Extensions.Assembly
 {
     using System.Reflection;
     using FluentAssertions;
-    using Tethos.Extensions;
+    using Tethos.Extensions.Assembly;
     using Xunit;
 
     public class FileExtensionsTests : BaseAutoMockingTest<AutoMockingContainer>
     {
         [Fact]
         [Trait("Category", "Unit")]
-        public void GetPath_FromExecutingAssembly_ShouldMatchLocation()
+        public void GetFile_FromExecutingAssembly_ShouldMatchLocation()
         {
             // Arrange
             var assembly = Assembly.GetExecutingAssembly();
